Integrate PluginInstaller into IdeaBasedIdeToolCommandlet
